+++ src/minimizers.jl
@@ -3,14 +3,21 @@
 #
 
 # --------------------------------------------------------------------
-@enum MinimizerStatusCode MinOK MinWARN MinERROR MinDRY
+abstract type AbstractMinimizerStatus end
 
-struct MinimizerStatus
-    code::MinimizerStatusCode
-    message::String
-    internal
+struct MinimizerStatusOK <: AbstractMinimizerStatus
 end
-MinimizerStatus(code::MinimizerStatusCode) = MinimizerStatus(code, "", nothing)
+
+struct MinimizerStatusDry <: AbstractMinimizerStatus
+end
+
+struct MinimizerStatusWarn <: AbstractMinimizerStatus
+    message::String
+end
+
+struct MinimizerStatusError <: AbstractMinimizerStatus
+    message::String
+end
 
 
 # --------------------------------------------------------------------
@@ -25,34 +32,37 @@ function fit(minimizer::dry, fp::AbstractFitProblem)
     finalize!(fp,
               getfield.(params, :val),
               fill(NaN, length(params)))
-    return MinimizerStatus(MinDRY)
+    return MinimizerStatusDry
 end
 
 
 # --------------------------------------------------------------------
 import LsqFit
 
-struct lsqfit <: AbstractMinimizer; end
+mutable struct lsqfit <: AbstractMinimizer;
+    result
+    lsqfit() = new(nothing)
+end
 
-function fit(minimizer::lsqfit, fp::AbstractFitProblem)
+function fit(mzer::lsqfit, fp::AbstractFitProblem)
     params = free_params(fp)
     ndata = length(residuals(fp))
     prog = ProgressUnknown(desc="Model (dof=$(fp.dof)) evaluations:", dt=0.5, showspeed=true, color=:light_black)
-    res = LsqFit.curve_fit((dummy, pvalues) -> begin
-                               ProgressMeter.next!(prog; showvalues=() -> [(:fit_stat, fit_stat(fp))])
-                               residuals(fp, pvalues)
-                           end,
-                           1.:ndata, fill(0., ndata),
-                           getfield.(params, :val),
-                           lower=getfield.(params, :low),
-                           upper=getfield.(params, :high))
+    mzer.result = LsqFit.curve_fit((dummy, pvalues) -> begin
+                                       ProgressMeter.next!(prog; showvalues=() -> [(:fit_stat, fit_stat(fp))])
+                                       residuals(fp, pvalues)
+                                   end,
+                                   1.:ndata, fill(0., ndata),
+                                   getfield.(params, :val),
+                                   lower=getfield.(params, :low),
+                                   upper=getfield.(params, :high))
     ProgressMeter.finish!(prog)
-    if !res.converged
-        return MinimizerStatus(MinERROR, "Not converged", res)
+    if !mzer.result.converged
+        return MnimizerStatusError("Not converged")
     end
 
-    finalize!(fp, getfield.(Ref(res), :param), LsqFit.stderror(res))
-    return MinimizerStatus(MinOK, "", res)
+    finalize!(fp, getfield.(Ref(mzer.result), :param), LsqFit.stderror(mzer.result))
+    return MinimizerStatusOK()
 end
 
 
@@ -75,14 +85,15 @@ iterated for the maximum allowed number of times.
 mutable struct cmpfit <: AbstractMinimizer
     config::CMPFit.Config
     ftol_after_maxiter::Float64
+    result
     function cmpfit()
-        out = new(CMPFit.Config(), 1e-4)
+        out = new(CMPFit.Config(), 1e-4, nothing)
         out.config.maxiter = 1000
         return out
     end
 end
 
-function fit(minimizer::cmpfit, fp::AbstractFitProblem)
+function fit(mzer::cmpfit, fp::AbstractFitProblem)
     params = free_params(fp)
     guess = getfield.(params, :val)
     low   = getfield.(params, :low)
@@ -99,35 +110,35 @@ function fit(minimizer::cmpfit, fp::AbstractFitProblem)
     last_fitstat = sum(abs2, residuals(fp))
     prog = ProgressUnknown(desc="Model (dof=$(fp.dof)) evaluations:", dt=0.5, showspeed=true, color=:light_black)
     while true
-        res = CMPFit.cmpfit((pvalues) -> begin
-                                ProgressMeter.next!(prog; showvalues=() -> [(:fit_stat, fit_stat(fp))])
-                                residuals(fp, pvalues)
-                            end,
-                            guess, parinfo=parinfo, config=minimizer.config)
-        if res.status <= 0
-            return MinimizerStatus(MinError, "Status = $(res.status)", res)
+        mzer.result = CMPFit.cmpfit((pvalues) -> begin
+                                        ProgressMeter.next!(prog; showvalues=() -> [(:fit_stat, fit_stat(fp))])
+                                        residuals(fp, pvalues)
+                                    end,
+                                    guess, parinfo=parinfo, config=mzer.config)
+        if mzer.result.status <= 0
+            return MinimizerStatusError("CMPFit status = $(mzer.result.status)")
         end
 
-        if (res.status == 5)
-            Δfitstat = (last_fitstat - res.bestnorm) / last_fitstat
-            if Δfitstat > minimizer.ftol_after_maxiter
-                println("Reached max. number of iteration but relative Δfitstat = $(Δfitstat) > $(minimizer.ftol_after_maxiter), continue minimization...\n")
-                last_fitstat = res.bestnorm
-                guess = getfield.(Ref(res), :param)
+        if (mzer.result.status == 5)
+            Δfitstat = (last_fitstat - mzer.result.bestnorm) / last_fitstat
+            if Δfitstat > mzer.ftol_after_maxiter
+                println("Reached max. number of iteration but relative Δfitstat = $(Δfitstat) > $(mzer.ftol_after_maxiter), continue minimization...\n")
+                last_fitstat = mzer.result.bestnorm
+                guess = getfield.(Ref(mzer.result), :param)
                 continue
             end
         end
 
         ProgressMeter.finish!(prog)
         finalize!(fp,
-                  getfield.(Ref(res), :param),
-                  getfield.(Ref(res), :perror))
+                  getfield.(Ref(mzer.result), :param),
+                  getfield.(Ref(mzer.result), :perror))
 
-        if res.status == 2
-            return MinimizerStatus(MinWARN, "CMPFit status = 2 may imply one (or more) guess values are too far from optimum", res)
-        elseif res.status == 5
-            return MinimizerStatus(MinWARN, "CMPFit status = 5, reached maximum allowed number of iteration.", res)
+        if mzer.result.status == 2
+            return MinimizerStatusWarn("CMPFit status = 2 may imply one (or more) guess values are too far from optimum")
+        elseif mzer.result.status == 5
+            return MinimizerStatusWarn("CMPFit status = 5, reached maximum allowed number of iteration.")
         end
-        return MinimizerStatus(MinOK, "", res)
+        return MinimizerStatusOK()
     end
 end